requirements.txt
langchain
wikipedia
langchain-community
langchain-google-genai
langchain-google-community
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
python-dotenv
pydantic
requests